+++ src/main/java/com/cloudogu/smeagol/wiki/usecase/FailedToInitWikiException.java
@@ -1,0 +1,15 @@
+package com.cloudogu.smeagol.wiki.usecase;
+
+/**
+ * Exception is thrown if failed to initialize a wiki.
+ */
+public class FailedToInitWikiException extends RuntimeException {
+
+    public FailedToInitWikiException(String message) {
+        super(message);
+    }
+
+    public FailedToInitWikiException(Throwable cause) {
+        super(cause);
+     }
+}